Repeat chunks where we didn't read in enough particles
Halos in each chunk task have a flag to indicate if they have been
processed. If a halo can't be processed because the region read in was
too small the read radius is doubled and we repeat the chunk, only
reading in particles near the remaining halos.

TODO: fix reports of task timing and number of halos processed.