Avoid gathering metadata from all chunks and MPI ranks
Each chunk task now checks that all ranks on the node agree on
the quantities computed. Only the first rank on the node returns
the list of quantities.

At the end of the calculation each node checks the results of the
chunks it processed for consistency. Then each node sends its list
of quantities to rank 0 in comm_world to check that all nodes
agree.